Sandbox room overhaul
New Sandbox room created. 3 inputs and 3 outputs.

The button causes the table to change it's output values based on the logic gate configuration in the boundary area. Resets every time the button is pressed.

Note: The script for table writing is limited to the single table and hard coded at the moment. Already planning to refactor this in the near future.